refactor(aten): share index helpers across index_select / index_copy
Extract the helpers that index_select and index_copy were each defining
locally into a single header so there is one source of truth:

  - read_index_to_host(index, op_name)  — pull a 0/1-D index tensor onto
    the host as int64 (was duplicated; op_name parameter keeps the
    RBLN_CHECK messages naming the caller).
  - struct IndexRun + coalesce_runs()   — maximal +1-increment runs.
    Renamed fields to op-neutral `value` (axis position) and `pos`
    (position in the index list); callers narrow accordingly.

`is_same_view(a, b)` was also duplicated — once as `is_identity_copy`
inside RBLNStridedV2V's anonymous namespace, once in RBLNIndexCopy.
Moved to RBLNTensorUtils.h as an inline helper so the engine and the
in-place dispatch detection share the same implementation.

No behaviour change; -123 net lines.
